remove galois.tar and index.html, rename README.html
remove galois.tar, quoting jplank@ "my guess is that the galois.tar
there was mistakenly tarred up with the source files — since that was a
lifetime ago, I don’t remember!  I would ignore it."

remove index.html.  it's a hard link to (or from) README.html
originally, but we no longer need it with the repo currently served up
on github.

rename README.html to README.md.  github won't parse markdown (or the
html subset) without it bearing the .md extension.